ref - prt - Savers config migrated to one-class
---

The extra <ScreensaverName>Settings static classes used to run the value checks while the kernel settings were intact before the whole kernel configuration system was changed to use the serialization and the deserialization technique.

These classes are now removed with this commit, but this is not a breaking change. This simplifies the value checks for every single screensaver.

---

Type: ref
Breaking: False
Doc Required: False
Part: 2/2

+++ public/Nitrocid.Addons/Nitrocid.ScreensaverPacks/Screensavers/Linotypo.cs
@@ -232,8 +232,8 @@ namespace Nitrocid.ScreensaverPacks.Screensavers
                             DebugWriter.WriteDebugConditional(ScreensaverManager.ScreensaverDebug, DebugLevel.I, "Etaoin type: {0}", ScreensaverPackInit.SaversConfig.LinotypoEtaoinType);
                             switch (ScreensaverPackInit.SaversConfig.LinotypoEtaoinType)
                             {
-                                case (int)FillType.EtaoinComplete:
-                                case (int)FillType.EtaoinPattern:
+                                case FillType.EtaoinComplete:
+                                case FillType.EtaoinPattern:
                                     {
                                         if (ScreensaverPackInit.SaversConfig.LinotypoEtaoinType == (int)FillType.EtaoinPattern)
                                             LinotypeMaxColumnIndex = 1;
@@ -255,7 +255,7 @@ namespace Nitrocid.ScreensaverPacks.Screensavers
 
                                         break;
                                     }
-                                case (int)FillType.RandomChars:
+                                case FillType.RandomChars:
                                     {
                                         // Randomly select the linotype indexes
                                         LinotypeColumnIndex = RandomDriver.Random(0, 5);
